requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
decorator==4.4.2
Flask==2.0.1
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.26.1
Pillow==10.1.0
proglog==0.1.10
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
Werkzeug==2.0.1
